feat: don't require next_frame() on first step
It should be considered invalid to have negative
frame ids. On the very first step, currently the
collector expects the user to increment the frame
once before starting to collect (leading-edge),
however it should be equally valid to increment
the frame after collecting (trailing-edge).

Adds a tiny amount of latency to the write path
since now we have to do a check for _frame=-1.