feat: serialize in memory dbs
Not a particularly memory efficient strategy,
we first dump the db into a string (semi-colon
separated set of sql commands to rebuild the db),
then store that string directly in the pickle.
To rebuild, we detect if the string exists and
build a new db from it. Quite inefficient, so
should add warning in documentation not to use
in-memory dbs for long running jobs.